refactoring plugins-avaliable component
