+++ src/app/pages/plugins/plugins.module.ts
@@ -7,7 +7,7 @@ import { MaterialModule, MdTableModule } from '@angular/material';
 import {EntityModule} from '../common/entity/entity.module';
 
 import {routing} from './plugins.routing';
-import { PluginsAvailabelListComponent } from './plugins-available/plugins-available-list/plugins-available-list.component';
+import { PluginsAvailabelListComponent } from './plugins-available/plugins-available-list.component';
 import { PluginAddComponent } from './plugin-add/plugin-add.component';
 import { PluginsInstalledListComponent } from './plugins-installed/plugins-installed.component';
