+++ src/app/pages/plugins/plugins.routing.ts
@@ -1,7 +1,7 @@
 import { ModuleWithProviders } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
-import { PluginsAvailabelListComponent } from './plugins-available/plugins-available-list/plugins-available-list.component';
+import { PluginsAvailabelListComponent } from './plugins-available/plugins-available-list.component';
 import { PluginAddComponent } from './plugin-add/plugin-add.component';
 import { PluginsInstalledListComponent } from './plugins-installed/plugins-installed.component';
